Her har vi nå satt slik at formen legges til i databasen, slik at vi kan se den i rapports tabellen i admin siden - Emil

+++ Models/GetAllReportsModel.cs
@@ -9,8 +9,8 @@
         public class MakeReportList
         {
             public string Navn { get; set; }
-            public int Lattitude { get; set; }
-            public int Longitude { get; set; }
+            public double Lattitude { get; set; }
+            public double Longitude { get; set; }
             public int Altitude { get; set; }
         }
     }
